Rsisolto errore gestione timer turno

+++ src/main/java/Progetto_Ing_Sw/com/client/SocketClient.java
@@ -130,7 +130,7 @@ public class SocketClient implements Runnable{
                 localModel.setCountdownValue(Long.parseLong(messageFields[1]));
                 break;
             case "Your turn will end in":
-                localModel.setTurnCountDownValue(Long.parseLong(messageFields[1]));
+                localModel.setTurnCountDownValue(Integer.parseInt(messageFields[1]));
                 break;
             case "Dice placed successfully":
                 System.out.println("Dice placed successfully");
